softlist wip: * added list for svi318 tapes and disks [Kaylee] * hooked up a7800 list (the code will need improvements later) [Tafoid, myself] * hooked up svi318 new lists
git-svn-id: svn://messdev.no-ip.org/mess@10198 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/includes/a7800.h
@@ -23,8 +23,8 @@ public:
 	int ctrl_reg;
 	int maria_flag;
 	unsigned char *cartridge_rom;
-	unsigned int cart_type;
-	unsigned long cart_size;
+	UINT16 cart_type;
+	UINT32 cart_size;
 	unsigned char stick_type;
 	UINT8 *ROM;
 	int maria_palette[8][4];
